+++ src/snow/object/grid.cpp
@@ -47,16 +47,16 @@ void Grid::debug() const {
   glUnmapBuffer(GL_SHADER_STORAGE_BUFFER);
 
   glBindBuffer(GL_SHADER_STORAGE_BUFFER, velB);
-  Vector4i* c = (Vector4i*)(glMapBufferRange(
-      GL_SHADER_STORAGE_BUFFER, 0, sizeof(Vector4i) * gridPoints.size(),
+  Vector4f* c = (Vector4f*)(glMapBufferRange(
+      GL_SHADER_STORAGE_BUFFER, 0, sizeof(Vector4f) * gridPoints.size(),
       GL_MAP_READ_BIT));
   std::cout << "vi: ";
   c[(9 + 5 * 201 + 9 * 201 * 201)].print();
   glUnmapBuffer(GL_SHADER_STORAGE_BUFFER);
 
   glBindBuffer(GL_SHADER_STORAGE_BUFFER, forceB);
-  Vector4i* f = (Vector4i*)(glMapBufferRange(
-      GL_SHADER_STORAGE_BUFFER, 0, sizeof(Vector4i) * gridPoints.size(),
+  Vector4f* f = (Vector4f*)(glMapBufferRange(
+      GL_SHADER_STORAGE_BUFFER, 0, sizeof(Vector4f) * gridPoints.size(),
       GL_MAP_READ_BIT));
   std::cout << "fi: ";
   f[(9 + 5 * 201 + 9 * 201 * 201)].print();
@@ -118,32 +118,32 @@ void Grid::initSSBO() {
 
   glGenBuffers(1, &velB);
   glBindBuffer(GL_SHADER_STORAGE_BUFFER, velB);
-  glBufferData(GL_SHADER_STORAGE_BUFFER, sizeof(Vector4i) * gridPoints.size(),
+  glBufferData(GL_SHADER_STORAGE_BUFFER, sizeof(Vector4f) * gridPoints.size(),
                NULL, GL_STATIC_DRAW);
   glUnmapBuffer(GL_SHADER_STORAGE_BUFFER);
   glBindBufferBase(GL_SHADER_STORAGE_BUFFER, GRID_VEL_BUFFER, velB);
   std::cout << "GridVelocityBufferSize: "
-            << sizeof(Vector4i) * gridPoints.size() / 1024 << " KB"
+            << sizeof(Vector4f) * gridPoints.size() / 1024 << " KB"
             << std::endl;
 
   glGenBuffers(1, &forceB);
   glBindBuffer(GL_SHADER_STORAGE_BUFFER, forceB);
-  glBufferData(GL_SHADER_STORAGE_BUFFER, sizeof(Vector4i) * gridPoints.size(),
+  glBufferData(GL_SHADER_STORAGE_BUFFER, sizeof(Vector4f) * gridPoints.size(),
                NULL, GL_STATIC_DRAW);
   glUnmapBuffer(GL_SHADER_STORAGE_BUFFER);
   glBindBufferBase(GL_SHADER_STORAGE_BUFFER, GRID_FORCE_BUFFER, forceB);
   std::cout << "GridForceBufferSize: "
-            << sizeof(Vector4i) * gridPoints.size() / 1024 << " KB"
+            << sizeof(Vector4f) * gridPoints.size() / 1024 << " KB"
             << std::endl;
 
   glGenBuffers(1, &velBn);
   glBindBuffer(GL_SHADER_STORAGE_BUFFER, velBn);
-  glBufferData(GL_SHADER_STORAGE_BUFFER, sizeof(Vector4i) * gridPoints.size(),
+  glBufferData(GL_SHADER_STORAGE_BUFFER, sizeof(Vector4f) * gridPoints.size(),
                NULL, GL_STATIC_DRAW);
   glUnmapBuffer(GL_SHADER_STORAGE_BUFFER);
   glBindBufferBase(GL_SHADER_STORAGE_BUFFER, GRID_VEL_N_BUFFER, velBn);
   std::cout << "GridVelocityBufferSize: "
-            << sizeof(Vector4i) * gridPoints.size() / 1024 << " KB"
+            << sizeof(Vector4f) * gridPoints.size() / 1024 << " KB"
             << std::endl;
 }
